feat: add landing page for game
